% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{add_additional_terms}
\alias{add_additional_terms}
\title{Add additional terms to the covariate equation}
\usage{
add_additional_terms(fes, add_terms = NULL)
}
\arguments{
\item{fes}{original equation}

\item{add_terms}{a vector of new terms to add to the covariate equation}
}
\value{
the covariate/fixed effects equation with additional terms tacked on
}
\description{
Add additional terms to the covariate equation
}
