% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{centreScale}
\alias{centreScale}
\title{Center (e.g. normalize) data}
\usage{
centreScale(x, df, inverse = FALSE)
}
\arguments{
\item{x}{data frame (or similiar). The data object}

\item{df}{data frame. data frame providing mean and standard deviations from getCentreScale}

\item{inverse}{logical. If T, x is unscaled rather than scaled/normalized}
}
\description{
Center (e.g. normalize) data
}
