% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{check_names}
\alias{check_names}
\title{Check names}
\usage{
check_names(obj, required_names)
}
\arguments{
\item{obj}{any object with a names attribute.}

\item{required_names}{character vector. The names an object must have to pass the test}
}
\description{
Ensure an object has certain names. Written as a test rather than something that returns T/F
}
