% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_submodels.R
\name{check_submodels}
\alias{check_submodels}
\title{A little wrapper to check the status of submodels}
\usage{
check_submodels(st)
}
\arguments{
\item{st}{stacking governer}
}
\description{
A little wrapper to check the status of submodels
}
