% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile_submodels.R
\name{compile_submodels}
\alias{compile_submodels}
\title{Compile submodels
Turns a list of results created by various fit* functions into a two item list. Item 1 is a data.table of full and cv predictions. Item 2 is a list of model objects}
\usage{
compile_submodels(st, stacking_models)
}
\arguments{
\item{st}{stacking governer}

\item{stacking_models}{list. List of results from the various fit* functions}
}
\description{
Compile submodels
Turns a list of results created by various fit* functions into a two item list. Item 1 is a data.table of full and cv predictions. Item 2 is a list of model objects
}
