% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{emplogit}
\alias{emplogit}
\title{Emperical logit}
\usage{
emplogit(success, N, epsilon = NULL)
}
\arguments{
\item{success}{numeric vector. Binomial successes}

\item{N}{numeric vector. Same length as success. Represents the number of binomial trials}

\item{epsilon}{numeric. Offset value for instances where success/N are equal to 0 or 1. If left NULL, the function defaults to 1/2 the minimum observed value}
}
\description{
Given two vectors (success and N), calculate the emperical logistc. This is used to deal with cases when success/N
equals 0 or 1.
}
\details{
This stack exchange explains the process: http://stats.stackexchange.com/questions/109702/empirical-logit-transformation-on-percentage-data
}
