% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_stacked_rasters.R
\name{fetch_covariate_layer}
\alias{fetch_covariate_layer}
\title{Make rasters from the child model objects}
\usage{
fetch_covariate_layer(ras, time_position = 1)
}
\arguments{
\item{ras}{raster-like. Raster like object}

\item{time_position}{numeric. Location on the time scale but reduced to units of 1:max(time_scale)}
}
\description{
Using model fits from run_stacking_child_models, creates rasters
}
