% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{find_binary}
\alias{find_binary}
\title{Find binary variables}
\usage{
find_binary(df)
}
\arguments{
\item{df}{A data table.}
}
\description{
Given a data table, returns the names of the columns that have two or fewer responses.
}
