% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_earth.R
\name{fit_earth}
\alias{fit_earth}
\title{Fit an Earth model}
\usage{
fit_earth(st, model_name = "earth", fold_col = NULL, fold_id = NULL,
  return_model_obj = F)
}
\arguments{
\item{st}{stacker governor. Stacking governer object with an earth model initialized. See mbgstacking::init_earth for more details.}

\item{model_name}{character vector. Name/model identifier of the model to be run}

\item{fold_col}{character vector. Denotes the name of the column designating the fold for crossvalidation}

\item{fold_id}{Numeric. Designates the value in fold col that should be held out}

\item{return_model_obj}{logical. Denotes whether the function should return the earth model object or just predictions.}
}
\value{
List object with a data.table of predictions. If return_model_obj==T, the earth command and model object are returned as well
}
\description{
Fit a multivariate regression spline model using earth::earth. See earth::earth for more details.
}
