% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gam_models.R
\name{fit_gam}
\alias{fit_gam}
\title{Fit gam model}
\usage{
fit_gam(st, model_name = "gam", fold_col = NULL, fold_id = NULL,
  return_model_obj = F)
}
\arguments{
\item{st}{stacker governer. Stacking governer object with an gam model initialized}

\item{model_name}{character vector. Name of the model to be run}

\item{fold_col}{character vector. Denotes the name of the column designating the fold for crossval}

\item{fold_id}{Numeric. Designates the value in fold col that should be held out}

\item{return_model_obj}{logical. Denotes whether the function should return the earth object or just predictions.}
}
\value{
List object with a data.table of predictions. If return_model_obj==T, the gam command and model object are returned as well
}
\description{
Fit a generalized additive model using mgcv::gam. See help(gam) for more details
}
