% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_glmnet.R
\name{fit_glmnet}
\alias{fit_glmnet}
\title{Fit penalized regression model}
\usage{
fit_glmnet(st, model_name = "pen", fold_col = NULL, fold_id = NULL,
  return_model_obj = F)
}
\arguments{
\item{st}{stacker governor. Stacking governer object with a penalized model initialized}

\item{model_name}{character vector. Name of the model to be run}

\item{fold_col}{character vector. Denotes the name of the column designating the fold for crossval}

\item{fold_id}{Numeric. Designates the value in fold col that should be held out}

\item{return_model_obj}{logical. Denotes whether the function should return the earth object or just predictions.}
}
\value{
List object with a data.table of predictions. If return_model_obj==T, the gam command and model object are returned as well
}
\description{
Fit a penalized regression model. See ??glmnet::glmnet for more details
}
