% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_xgb.train.R
\name{fit_xgb.train}
\alias{fit_xgb.train}
\title{Fit boosted regression tree/gradient boosted machine}
\usage{
fit_xgb.train(st, model_name = "brt", fold_col = NULL, fold_id = NULL,
  return_model_obj = F, sub_cores = 1)
}
\arguments{
\item{st}{stacker governor. Stacking governer object with a penalized model initialized}

\item{model_name}{character vector. Name of the model to be run}

\item{fold_col}{character vector. Denotes the name of the column designating the fold for crossval}

\item{fold_id}{Numeric. Designates the value in fold col that should be held out}

\item{return_model_obj}{logical. Denotes whether the function should return the earth object or just predictions.}

\item{sub_cores}{numeric. Number of cores/processes to be used by xgboost.}
}
\value{
List object with a data.table of predictions. If return_model_obj==T, the gam command and model object are returned as well
}
\description{
Fit a boosted regression tree (or random forest given the proper set of options). See ??xgboost::xgb.train for more details
}
