% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{format_covariates}
\alias{format_covariates}
\title{Format covariate equation}
\usage{
format_covariates(covariates)
}
\arguments{
\item{covariates}{a formula-like string}
}
\value{
A vector where each item is a term in the formula
}
\description{
Splits a formula like string into a vector where each item is a term
}
