% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{format_model_covs}
\alias{format_model_covs}
\title{Format Model Covariates}
\usage{
format_model_covs(covariates, additional_terms)
}
\arguments{
\item{covariates}{a formula-like string}

\item{additional_terms}{a string vector of new terms}
}
\value{
A vector where each item is a term in the formula
}
\description{
Returns the covariate equation as a vector with each item as a term of the equation with additional terms appended on
}
