% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{get_model_type}
\alias{get_model_type}
\title{Get Model Type}
\usage{
get_model_type(st, model_name)
}
\arguments{
\item{st}{stacker object}

\item{model_name}{character string of the model name}
}
\value{
A string with the model type
}
\description{
Returns the model type/class from the stacker object given a model name
}
