% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{getCentreScale}
\alias{getCentreScale}
\title{Center (e.g. normalize) data}
\usage{
getCentreScale(x, exclude = NULL, na.rm = TRUE)
}
\arguments{
\item{x}{data frame.}

\item{exclude}{character vector. List of columns in x to ignore (e.g. leave unstandardized)}

\item{na.rm}{logical. na.rm setting for colmeans and sd.}
}
\description{
Center (e.g. normalize) data
}
