% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_models.R
\name{init_brt}
\alias{init_brt}
\title{Initialize a brt model}
\usage{
init_brt(model_name = "brt", arguments = list(), params_arg = list(nthread
  = 1), nrounds = 10, emp_logit = F)
}
\arguments{
\item{model_name}{name of the brt model}

\item{arguments}{named list. Arguments to be passed to the xgboost::xgb.train function.
Should only be used for options not captured by the params argument in the underlying xgboost function call.
See help xgboost::xgb.train for more information}

\item{params_arg}{named list. Arguments to be passed to the parameters argument of the xgboost::xgb.train function}

\item{nrounds}{numeric. Max number of iterations}

\item{emp_logit}{logical. Whether a binomial model should be be transformed using an emperical logit function instead of using the poisson approximation.
xgboost's binomial approach is exclusive to 1s and 0s}
}
\value{
named list of lists with the parameters required to run an brt model
}
\description{
Creates an object designed to be passed to init_stacker that describes an xgboost::xgb.train model be added to the stacking ensemble.
}
