% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_models.R
\name{init_earth}
\alias{init_earth}
\title{Initialize an earth model}
\usage{
init_earth(model_name = "earth", arguments = list(degree = NULL, nk = NULL))
}
\arguments{
\item{model_name}{name of the earth model}

\item{arguments}{named list. Arguments to be passed to the earth function. See ??earth::earth for more information}
}
\value{
named list of lists with the parameters required to run an earth model
}
\description{
Creates an object designed to be passed to init_stacker that describes an earth::earth model be added to the stacking ensemble.
}
