% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_models.R
\name{init_gam}
\alias{init_gam}
\title{Initialize a gam model}
\usage{
init_gam(model_name = "gam", arguments = list(spline_args = list(bs = "ts",
  k = 3)), formula = NULL)
}
\arguments{
\item{model_name}{name of the gam model}

\item{arguments}{named list. Arguments to be passed to the gan function. See help mgcv::gam for more information}

\item{formula}{formula or formula-like. Formula to be passed to the gam call that overwrites the underlying formula builder.
This is useful if you want to have different covariates recieve different smoothing instructions}
}
\value{
named list of lists with the parameters required to run an gam model
}
\description{
Creates an object designed to be passed to init_stacker that describes an gam model be added to the stacking ensemble.
}
