% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_penalized.R
\name{init_penalized}
\alias{init_penalized}
\title{Initialize a penalized linear regression model}
\usage{
init_penalized(model_name = "pen", arguments = list(alpha = 1),
  emp_logit = F, standardize = F)
}
\arguments{
\item{model_name}{name of the penalized regression model model}

\item{arguments}{named list. Arguments to be passed to the glmnet function. See help glmnet::glmnet for more information.
The main arguement to be passed is alpha: 0 is ridge regression and 1 is lasso penalty. Between 0 and 1 refers to elastic net.}

\item{emp_logit}{logical. If family is binomial, should the regression be run as gaussian (empirical logit of cases/N)}

\item{standardize}{logical. Standardize numeric columns to have zero mean and unit variance. Defaults to False unlike the glmnet default settings.
This is set to F, because centre scaling/normalizing is a default preprocessing step}
}
\value{
named list of lists with the parameters required to run an penalized regression model
}
\description{
Creates an object designed to be passed to init_stacker that describes an glm/penalized regression model be added to the stacking ensemble.
}
