% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_models.R
\name{init_sge}
\alias{init_sge}
\title{Initialize SGE parameters}
\usage{
init_sge(working_folder, r_path, output_files = NULL, error_files = NULL,
  project_name = NULL, other_options = NULL, slots_per_job = c(3, 5),
  package_location = NULL, repeat_iterations = 0, conda_activate = NULL,
  conda_env = NULL, write_shell = F)
}
\arguments{
\item{working_folder}{file path. Location accessible to the computing cluster where scratch files can be saved}

\item{r_path}{file path. Full path to R}

\item{output_files}{character string. Location where the output (.o) files are saved}

\item{error_files}{character string. Location where the error (.e) files are save}

\item{project_name}{character string. Cluster project}

\item{other_options}{character string. Character string of additional options to pass to qsub}

\item{slots_per_job}{numeric. Denotes the number of slots to be requested by each submodel. If length>1, first item is passed to child models and second to raster creation}

\item{package_location}{character string. Denotes the location where mbgstacking is installed (to be passed to the qsubs)}

\item{repeat_iterations}{numeric: How many times should jobs be relaunched if they don't work the first time.}

\item{conda_activate}{file path. Denotes the location where a conda environment activate option is. Not implemented.}

\item{conda_env}{file path. File path to the name of the conda environment. Not implemented.}

\item{write_shell}{logical. Should the qsub be run via writing shell scripts. If F, qsubs are lanunched via the "-b y" flag}
}
\value{
List of lists containing the input parameters to be passed to the stacker
}
\description{
Creates an object designed to be passed to init_stacker that describes how the internal machinery should handle .
}
