% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_models.R
\name{init_stacker}
\alias{init_stacker}
\title{Initialize stacker}
\usage{
init_stacker(..., data, indicator, indicator_family, covariate_layers,
  fe_equation, centre_scale = T, time_var = "year", time_scale = c(2000,
  2005, 2010, 2015), weight_col = NULL, num_fold_cols = 1, num_folds = 1,
  cores = 1, sge_parameters = NULL)
}
\arguments{
\item{...}{Initialized models. If blank, default versions of earth and gam are created.}

\item{data}{data table. Dataset to be machine learned.}

\item{indicator}{character vector. Name of the indicator (and by extension) the column name of the dependant variable}

\item{indicator_family}{character vector. Designates the statistical family that should be modeled. Usually 'binomial' or 'gaussian'}

\item{covariate_layers}{list of raster like objects. A named list of raster like objects of covariates}

\item{fe_equation}{character vector of an equation. The equation specifying the fixed effects portion of the model. It should match with the names
of covariate_layers.}

\item{centre_scale}{logical. Determines whether the covariate values are centered/normalized before being returned. Binary variables are ignored.}

\item{time_var}{character vector. Name of the column denoting the time (e.g. period or year) of a given data point}

\item{time_scale}{numeric vector. List of years or times that the time var correlates to.}

\item{weight_col}{character vector. Denotes the column (if applicable) in the dataset that specifies the data weights}

\item{num_fold_cols}{numeric or character. Number of columns/interations for crossfold validation. if a character string, assume it refers to columns already existing in data.
They will be renamed to sfold_#}

\item{num_folds}{numeric. The number of folds the data is split on.}

\item{cores}{numeric. The number of cores available for parallel computation}

\item{sge_parameters}{object returned from init_sge. Provides sge parameters to govern submodel computation. If NULL, mclapply is used to run submodels instead}
}
\value{
Stacker governor object
}
\description{
Creates a stacker governor containing the various information and data to run stacked generalization in IHME's MBG framework
}
