% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_gbm3.R, R/init_xgboost.R
\name{init_xgboost}
\alias{init_xgboost}
\alias{init_xgboost}
\title{Initialize a GBM3 brt model}
\usage{
init_xgboost(model_name = "xgboost", arguments = list(),
  params_arg = list(nthread = 1), nrounds = 10,
  binomial_evaluation = "prev", weight_column = "")

init_xgboost(model_name = "xgboost", arguments = list(),
  params_arg = list(nthread = 1), nrounds = 10,
  binomial_evaluation = "prev", weight_column = "")
}
\arguments{
\item{model_name}{name of the brt model}

\item{arguments}{named list. Arguments to be passed to the gbm3::gbmt_fit function.
Should only be used for options not captured by the params argument in the underlying xgboost function call.
See help xgboost::xgb.train for more information}

\item{params_arg}{named list. Arguments to be passed to the parameters argument of the xgboost::xgb.train function}

\item{nrounds}{numeric. Max number of iterations}

\item{binomial_evaluation}{one of 'prev', 'poisson', or 'emplogit'. Prev fits on indicator/N with reg:logistic evaluation.
Poisson uses log(N) as an offset while modelling under a poisson. Emplogit transforms things and runs under gaussian family}

\item{training_params}{named list. Arguments to be passed to the gbm3::training_params function}

\item{model_name}{name of the brt model}

\item{arguments}{named list. Arguments to be passed to the xgboost::xgb.train function.
Should only be used for options not captured by the params argument in the underlying xgboost function call.
See help xgboost::xgb.train for more information}

\item{nrounds}{numeric. Max number of iterations}

\item{binomial_evaluation}{one of 'prev', 'poisson', or 'emplogit'. Prev fits on indicator/N with reg:logistic evaluation.
Poisson uses log(N) as an offset while modelling under a poisson. Emplogit transforms things and runs under gaussian family}
}
\value{
named list of lists with the parameters required to run an brt model

named list of lists with the parameters required to run an brt model
}
\description{
Creates an object designed to be passed to init_stacker that describes an gbm3::gbmt_fit model be added to the stacking ensemble.

Creates an object designed to be passed to init_stacker that describes an xgboost::xgb.train model be added to the stacking ensemble.
}
