% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{invlogit}
\alias{invlogit}
\title{Inverse logit transformation}
\usage{
invlogit(x)
}
\arguments{
\item{x}{numeric vector.}
}
\description{
Given a numeric vector, the function applies the inverse logit transformation: exp(x)/(1+exp(x))
}
