% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{logit}
\alias{logit}
\title{Logit transformation}
\usage{
logit(x)
}
\arguments{
\item{x}{numeric vector.}
}
\description{
Given a numeric vector, the function applies the logit transformation: log(x/(1-x))
}
