% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_stacked_rasters.R
\name{make_all_children_rasters}
\alias{make_all_children_rasters}
\title{Make rasters from the child model objects}
\usage{
make_all_children_rasters(st, model_objects, time_points = NULL)
}
\arguments{
\item{st}{stacker governer.}

\item{model_objects}{a named list of model objects}

\item{time_points}{numeric vector. the name/identifier/subset to be predicted-- on the scale of time_var}
}
\description{
Using model fits from run_stacking_child_models, creates rasters
}
