% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{make_model_grid}
\alias{make_model_grid}
\title{Generate a model grid from a stacker object}
\usage{
make_model_grid(st, add_parents = T)
}
\arguments{
\item{st}{stacker governor.}

\item{add_parents}{logical. Add the "parent" models (e.g. the full ones where we want the model object)}
}
\description{
Generate a model grid from a stacker object
}
