% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{make_random_cov}
\alias{make_random_cov}
\title{Make random covariate
  Makes a raster from a random field}
\usage{
make_random_cov(template, nu, sd, scale)
}
\description{
Make random covariate
  Makes a raster from a random field
}
