% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{make_random_tvcov}
\alias{make_random_tvcov}
\title{Make timevarying covariate}
\usage{
make_random_tvcov(start_point, nperiods)
}
\arguments{
\item{start_point}{raster. A raster that is the starting point for a projection into nperiods}

\item{nperiods}{numeric. The number of layers the resulting brick should have. Alternatively, the number of
years the covariate should represent.}
}
\value{
A rasterbrick with `nperiods` layers projected forward from some starting point
}
\description{
Make timevarying covariate
}
