% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{make_stacking_folds}
\alias{make_stacking_folds}
\title{Make stacking folds}
\usage{
make_stacking_folds(numrows, numfolds = 8, numsets = 1)
}
\arguments{
\item{numrows}{numeric. The number of data points/rows in the dataset}

\item{numfolds}{numeric. The number of folds within a set}

\item{numsets}{numeric. The number of fold sets to return}
}
\description{
Assigns rows to different folds to support the cross validation inherent in stacking.
}
