% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{make_test_train}
\alias{make_test_train}
\title{Make Test and Train identifiers}
\usage{
make_test_train(data, fold_col = NULL, fold_id = NULL)
}
\arguments{
\item{data}{data.table. Dataset to be split into test and train. Requires a column called 'rid' which is the row id.}

\item{fold_col}{character vector. Name of a column in data from which the test/train split should be made}

\item{fold_id}{numeric vector. Value designating which rows (given fold_col) should be left out.}
}
\description{
Assigns rows to test/train designations depending on fold_col/fold_id combination. If null, test and train are the entire dataset.
}
