% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{parseArgsS}
\alias{parseArgsS}
\title{Parse spline arguments}
\usage{
parseArgsS(l)
}
\arguments{
\item{l}{named list. A named list that when deconstructed can be used to tell mgcv:gam what to do}
}
\description{
Assigns rows to different folds to support the cross validation inherent in stacking.
}
