% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_model.R
\name{run_model}
\alias{run_model}
\title{A little wrapper to run models via SGE}
\usage{
run_model(st, task_id)
}
\arguments{
\item{st}{stacking governer}

\item{task_id}{integer. refers to row in st$model_grid}
}
\description{
A little wrapper to run models via SGE
}
