% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_child_models.R
\name{run_stacking_child_models}
\alias{run_stacking_child_models}
\title{Run child models for stacking}
\usage{
run_stacking_child_models(st)
}
\arguments{
\item{st}{stacker governer.}
}
\description{
Fits all models specified in the passed stacking object. M*C*K + M models are run where M is the number of initialized models,
C is the number of columns/iterations for crossfold validation and K is the number of folds.
}
