% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{sanitize_parameters}
\alias{sanitize_parameters}
\title{Sanitize paramers}
\usage{
sanitize_parameters(x)
}
\arguments{
\item{x}{a subset of the stacker object}
}
\value{
subset of the stacker object (e.g. list of lists) with no null entries
}
\description{
Given a list of lists (e.g. parameters guiding a stacker child model), remove all the NULL entries
}
