% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sge_compile_child_models.R
\name{sge_compile_child_models}
\alias{sge_compile_child_models}
\title{Compile the results of a child model}
\usage{
sge_compile_child_models(st, add_parents = T)
}
\arguments{
\item{st}{stacker governor.}

\item{add_parents}{logical. Should the function also collect the parent/full models?}
}
\value{
character string. Name of the sge launched
}
\description{
Compile the results of a child model
}
