% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{sge_hold_via_sync}
\alias{sge_hold_via_sync}
\title{Create a blank job to pause R until the child jobs complete}
\usage{
sge_hold_via_sync(st, job_name = "holder", hold_on_jobs)
}
\arguments{
\item{st}{stacker governor. Must have relevant sge settings}

\item{job_name}{character. Name of the holding job}

\item{hold_on_jobs}{character vector. The name of the jobs that this job should hold on.}
}
\description{
Create a blank job to pause R until the child jobs complete
}
