% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sge_run_child_model.R
\name{sge_run_child_model}
\alias{sge_run_child_model}
\title{Run a child model of the stacker using Sun Grid Engine}
\usage{
sge_run_child_model(st)
}
\arguments{
\item{st}{stacker governer object}
}
\value{
character string. Name of the sge launched
}
\description{
Run a child model of the stacker using Sun Grid Engine
}
