% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sge_run_make_child_raster.R
\name{sge_run_make_child_raster}
\alias{sge_run_make_child_raster}
\title{Run a child model of the stacker using Sun Grid Engine}
\usage{
sge_run_make_child_raster(st, model_obj_name, time_position)
}
\arguments{
\item{st}{stacker governer object}

\item{model_obj_name}{character. The model object name}

\item{time_position}{numeric. relative position within the time scale that a raster should be generated}
}
\value{
character string. Name of the sge launched
}
\description{
Run a child model of the stacker using Sun Grid Engine
}
