% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{simulate_stacking_data}
\alias{simulate_stacking_data}
\title{Simulate data for unit testing}
\usage{
simulate_stacking_data(seed = 1, n_ntv = 2, n_tv = 5, n_cells = 10000,
  n_datapoints = length(years) * 100, years = 2000:2016,
  year_weights = rep.int(1/length(years), length(years)),
  family = "gaussian", fraction_weighted = 0.2)
}
\arguments{
\item{seed}{numeric. to be a random seed}

\item{n_tv}{numeric. number of time varying covariates to include}

\item{n_cells}{numeric. number of cells in the underlying rasters. sqrt(n_cells) must return an integer}

\item{n_datapoints}{numeric. number of datapoints to return}

\item{years}{numeric. vector of the years to be represented}

\item{year_weights}{numeric. Vector of weights (with length(year_weights) == to length(years)) for stratified sampling of datapoints.
NULL implies no year based stratification, while the value implies the number of data points (e.g. weight * n_datapoints) for a given year.}

\item{family}{character. Statistical family of the output. Currently only biomial and gaussian are supported}

\item{fraction_weighted}{numeric. Specifies the fraction of the dataset should have a weight 0<x<1}

\item{n_ntc}{numeric. number of non time varying covariates}
}
\value{
a data set and corrosponding rasters
}
\description{
Simulate data to test stacking functions
}
