% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_models.R
\name{init_sge}
\alias{init_sge}
\title{Initialize SGE parameters}
\usage{
init_sge(working_folder, rscript_path, output_files = NULL,
  error_files = NULL, project_name = NULL, other_options = NULL,
  slots_per_job = 2, package_location = NULL, conda_activate = NULL,
  conda_env = NULL)
}
\arguments{
\item{working_folder}{file path. Location accessible to the computing cluster where scratch files can be saved}

\item{rscript_path}{file path. Full path to rscript.}

\item{output_files}{character string. Location where the output (.o) files are saved}

\item{error_files}{character string. Location where the error (.e) files are save}

\item{project_name}{character string. Cluster project}

\item{other_options}{character string. Character string of additional options to pass to qsub}

\item{slots_per_job}{numeric. Denotes the number of slots to be requested by each submodel. Leave null for automatic calculation (based on dataset size)}

\item{package_location}{character string. Denotes the location where mbgstacking is installed (to be passed to the qsubs)}

\item{conda_env}{file path. File path to the name of the conda environment}

\item{conda_source}{file path. Denotes the location where a conda environment activate option is}
}
\value{
List of lists containing the input parameters to be passed to the stacker
}
\description{
Creates an object designed to be passed to init_stacker that describes how the internal machinery should handle .
}
